requirements.txt
llama-cpp-python>=0.2.0
tkinter
numpy
pathlib
typing
ttkthemes
PyPDF2
python-docx
beautifulsoup4
json